Add command-timeout option
Allow commands to have a timeout, set in minutes. This is considered an emergency mechanism to stop CI being halted indefinitely. If a test inherently needs to pass or fail based on a timeout, this should be scripted into the command string instead.

It can be set for a command resource, like:

```
  describe command('sleep 100', timeout: 1) do
    its('exit_status') { should cmp 0 }
  end
```

It can also be set as a new option `command-timeout`.
A timeout set on a `command` resource takes precedence over the `command-timeout` option.

If no timeout is set at all, a command will default to timing out after 1 hour

Signed-off-by: James Stocks <jstocks@chef.io>