Remove the kitchen.chef.yml file
This is incredibly dated and doesn't look like it was ever setup correctly. It's mostly just logic copied from a kitchen.yml file in an omnibus install, but this has nothing to do with omnibus. Windows / mac tests with kitchen should be done in Azure Devops Pipelines similar to how we do it with chef to avoid the reliance on a developer locally running an hour of tests.

Signed-off-by: Tim Smith <tsmith@chef.io>