Allow for rspec 3.9 in the gemspec
rspec 3.10 is causing issues with chef/chef tests due to changes in
rspec-mock that we're working to resolve. Right now we'd like to get the
latest inspec releases into chef/chef w/o the need for rspec 3.10. At
runtime inspec doesn't need 3.10, but it does need it for tests, so
let's make sure we bring it in via the gemfile, but allow the older 3.9
release at runtime.

Signed-off-by: Tim Smith <tsmith@chef.io>